Add files via upload
The aim of this script is to process data from shapefiles and rasters. The focus is on converting a shapefile representing events in the ACLED area into a raster, with the added bonus of creating dynamic maps. More specifically, the process involves transforming the shapefile into a raster, and calculating the number of events in each cell corresponding to an area of 10 km².